Refactor Form container and reducer
