+++ src/redux/rootReducer.js
@@ -1,7 +1,7 @@
 import items from '../reducers/items-reducer'
 import table from '../reducers/table-reducer'
 import filter from 'toolBar/reducer'
-import form from '../reducers/form-reducer'
+import form from 'table/reducer'
 
 const rootReducer = (state = {}, action) => {
   const { selectionRange, selectionFillHandleRange } = state.table || {}
